\alias{bbread}
\name{The Brown Bread Data}
\docType{data}
\title{
The Brown Bread Data Set
}
\description{
The brown bread data set is composed of 570 panelists and 16 products, labelled A,...,P to protect the identity of the manufacturer. Each panelist rated a subset of 6 products using the 9-point Hedonic Scale.
}
\usage{data(bbread)}
\examples{
data(bbread) # Loads the brown bread data set
head(bbread) # Displays the first six rows of the brown bread data set
}
\keyword{ datasets }
\keyword{ Brown Bread Data }
\keyword{ Sensory Informed Designs }