\alias{iiris}
\name{The Iris Data with Missing Values}
\docType{data}

\title{
The Iris Data with Missing Values
}

\description{
Fisher's Irises with one measurement removed per observation. For details load data(iris,package="gclus").
}

\usage{data(iiris)}

\examples{
data(iiris) # Loads the modified Iris data.
head(iiris) # Displays the first six rows of the modified Iris Data
}
\keyword{datasets}
\keyword{Fisher's Irises}