\name{sensory}
\alias{sensory}
\docType{package}
\title{Cluster-Wise Imputation via a PEM Algorithm
}
\description{Contains the function CUUimpute() which performs model-based clustering and imputation simultaneously.
}
\details{
\tabular{ll}{
Package: \tab sensory\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-07\cr
License: \tab GPL (>=2) \cr
}
This package contains two data sets collected at Compusense Inc. in Guelph, Ontario, Canada, a modified version of Fisher's Irises, and the function CUUImpute().
}
\author{
Brian C. Franczak, Ryan P. Browne and Paul D. McNicholas \cr
Maintainer: Brian C. Franczak <bfrancza@math.mcmaster.ca>%%  ~~who you are~~
}
\references{
P.D. McNicholas and T.B. Murphy (2008). Parsimonious Gaussian Mixture Models. Statistics and Computing (18), 285-296.\cr
Browne, R.P., P.D. McNicholas, and C.J. Findlay (2013).  A partial EM algorithm for clustering white breads. arXiv preprint arXiv:1302.6625.
}
\keyword{ Clustering }
\keyword{ Imputation }
\keyword{ Expectation Maximization (EM) Algorithm }
\keyword{ Parsimonious Gaussian Mixture Models }
\keyword{ CUU Model }
\keyword{ Sensory Informed Designs }

\seealso{Details, examples, and references are given under \link{CUUimpute}}