\name{The White Bread Data}
\alias{wbread}
\docType{data}
\title{
The White Bread Data
}
\description{
The white bread data is a liking study composed of 420 panelists and 12 products, labelled A,...,L to protect the identity of the manufacturer. Each panelist rated 6 white breads using the 9-point Hedonic scale.
}
\usage{data(wbread)}

\references{
Browne, R.P., P.D. McNicholas, and C.J. Findlay (2013).  A partial EM algorithm for clustering white breads. arXiv prerpeint arXiv:1302.6625.
}
\examples{
data(wbread) # Loads the white bread data
head(wbread) # Displays the first six rows of the white bread data
}
\keyword{ datasets }
\keyword{ White Bread Data }
\keyword{ Sensory Informed Designs }